# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
...
  user.Token:
    properties:
      created:
        type: string
      id:
        type: integer
    type: object
...
  v1.TokenResponse:
    properties:
      token:
        type: string
    type: object
...
  web.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact:
    email: hello@vikunja.io
    name: General Vikunja contact
    url: http://vikunja.io/en/contact/
  description: |-
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Rights
    All endpoints which return a single item (list, task, namespace, etc.) - no array - will also return a `x-max-right` header with the max right the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the rights the user has.
    # Authorization
    **JWT-Auth:** Main authorization method, used for most of the requests. Needs `Authorization: Bearer <jwt-token>`-header to authenticate successfully.

    **BasicAuth:** Only used when requesting tasks via caldav.
    <!-- ReDoc-Inject: <security-definitions> -->
  license:
    name: GPLv3
    url: http://code.vikunja.io/api/src/branch/master/LICENSE
  title: Vikunja API
paths:
...
  /user/settings/token/caldav:
    get:
      consumes:
      - application/json
      description: Return the IDs and created dates of all caldav tokens for the current
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.Token'
            type: array
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Returns the caldav tokens for the current user
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Generates a caldav token which can be used for the caldav api.
        It is not possible to see the token again after it was generated.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TokenResponse'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Generate a caldav token
      tags:
      - user
  /user/settings/token/caldav/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Delete a caldav token by id
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"